Eliminate `cookies` package.
