+++ components/api/src/main/java/com/hotels/styx/api/RequestCookie.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.cookies;
+package com.hotels.styx.api;
 
 
 import io.netty.handler.codec.http.cookie.ClientCookieEncoder;
