+++ components/api/src/test/java/com/hotels/styx/api/ServerCookieEncoderTest.java
@@ -13,8 +13,9 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.cookies;
+package com.hotels.styx.api;
 
+import com.hotels.styx.api.ServerCookieEncoder;
 import io.netty.handler.codec.http.cookie.ClientCookieDecoder;
 import io.netty.handler.codec.http.cookie.Cookie;
 import org.testng.annotations.Test;
