+++ components/server/src/test/java/com/hotels/styx/server/netty/connectors/HttpResponseWriterTest.java
@@ -42,7 +42,7 @@ import java.util.concurrent.atomic.AtomicLong;
 import static com.google.common.base.Charsets.UTF_8;
 import static com.hotels.styx.api.HttpResponse.response;
 import static com.hotels.styx.api.StyxInternalObservables.fromRxObservable;
-import static com.hotels.styx.api.cookies.ResponseCookie.responseCookie;
+import static com.hotels.styx.api.ResponseCookie.responseCookie;
 import static com.hotels.styx.api.HttpResponseStatus.OK;
 import static io.netty.buffer.Unpooled.copiedBuffer;
 import static io.netty.handler.codec.http.LastHttpContent.EMPTY_LAST_CONTENT;
